daschlab/photometry.py: tidy up this split
We are splitting most of the lightcurve functionality into a new base
class, because we want to add support for "extracts", which will be
photometric catalogs extracted from regions of individual plates. A lot of
the infrastructure is generic as to whether a collection of photometry
is time-oriented (lightcurve) or space-oriented (extract).