Require the latest pywwt and toasty
The newest pywwt has a workaround for an issue in reproject that causes
it to fail with integer-valued images that have the BLANK keyword. This
prevents us from displaying many cutouts.

The newest Toasty adds support for multiple WCS keys, so that you can
choose which WCS solution to use to position a multi-exposure mosaic.
This isn't strictly required by anything in daschlab, but pywwt induces
the requirement.