daschlab/apiclient.py: describe the web API that failed, if that's the case
For the HD 5501 sample, I'm getting consistent timeouts in queryexps.
Unfortunately this change doesn't do anything to help make it clear
what's happening there.

For posterity: I've increased the timeout of the underlying Lambda, and
the API Gateway integration talking to that Lambda, but it appears that
we're still limited by a 30-second timeout enforced by the CloudFront
instance that is implicitly included in the processing stack because we
are using an "edge-optimized" API Gateway configuration. If we switch to
"regional", the CloudFront layer should go away and the timeouts might
go away. But it's 10 PM and I don't want to get mired into making a
change like that right now, so it'll have to wait.